Add scheduled macroeconomic refresh:
- Introduce `FredService` for handling FRED-based macroeconomic data ingestion.
- Extend `RefreshScheduler` with a new scheduled task for macroeconomic data updates.
- Add `FetchMacroeconomicRequestEvent` for event-driven refresh handling.
- Update `HarvesterBatchService` to broadcast macroeconomic and news refresh events.
- Refactor database schema and DTOs to handle `Double` for macroeconomic values, replacing `BigDecimal`.
- Add utility class `NumericUtil` for string-to-double conversion of FRED data.
- Introduce `macroeconomic` cron configuration in application properties.

+++ src/main/java/com/oraculum/economy/api/dto/MacroObservationDto.java
@@ -2,13 +2,12 @@ package com.oraculum.economy.api.dto;
 
 import com.oraculum.economy.api.domain.MacroIndicator;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record MacroObservationDto(
         MacroIndicator indicatorCode,
         LocalDate observationDate,
-        BigDecimal value
+        Double value
 ) {
     public MacroObservationKey getKey() {
         return new MacroObservationKey(indicatorCode, observationDate);
